requirements.txt
fastapi
uvicorn
tensorflow
pillow
pydantic
numpy
python-multipart